Organize utility classes.
