+++ app/src/main/java/me/devsaki/hentoid/components/ImageDownloadBatch.java
@@ -1,6 +1,5 @@
 package me.devsaki.hentoid.components;
 
-import java.io.File;
 import java.util.ArrayList;
 import java.util.List;
 import java.util.concurrent.ExecutorCompletionService;
