+++ app/src/main/java/me/devsaki/hentoid/util/Helper.java
@@ -1,25 +1,17 @@
 package me.devsaki.hentoid.util;
 
-import android.content.Context;
-import android.content.SharedPreferences;
-import android.os.Environment;
-
 import com.google.gson.Gson;
 import com.google.gson.GsonBuilder;
 
 import java.io.BufferedReader;
 import java.io.File;
-import java.io.FileFilter;
 import java.io.FileReader;
 import java.io.FileWriter;
 import java.io.IOException;
 
-import me.devsaki.hentoid.HentoidApplication;
-import me.devsaki.hentoid.database.domains.Content;
-import me.devsaki.hentoid.database.enums.Site;
-
 /**
  * Created by DevSaki on 10/05/2015.
+ * Generic utility class
  */
 public final class Helper {
 
@@ -30,98 +22,6 @@ public final class Helper {
         while (n-- > 0 && pos != -1)
             pos = str.indexOf(delimiter, pos + 1);
         return pos;
-    }
-
-    public static String getSessionCookie() {
-        return HentoidApplication.getAppPreferences()
-                .getString(ConstantsPreferences.WEB_SESSION_COOKIE, "");
-    }
-
-    public static void setSessionCookie(String sessionCookie) {
-        HentoidApplication.getAppPreferences()
-                .edit()
-                .putString(ConstantsPreferences.WEB_SESSION_COOKIE, sessionCookie)
-                .apply();
-    }
-
-    public static File getThumb(Content content, Context context) {
-        File dir = Helper.getDownloadDir(content, context);
-
-        File[] fileList = dir.listFiles(
-                new FileFilter() {
-                    @Override
-                    public boolean accept(File pathname) {
-                        return pathname.getName().contains("thumb");
-                    }
-                }
-        );
-
-        return fileList.length > 0 ? fileList[0] : null;
-    }
-
-    public static File getDownloadDir(Content content, Context context) {
-        File file;
-        SharedPreferences prefs = HentoidApplication.getAppPreferences();
-        String settingDir = prefs.getString(Constants.SETTINGS_FOLDER, "");
-        String folderDir = content.getSite().getFolder() + content.getUniqueSiteId();
-        if (settingDir.isEmpty()) {
-            return getDefaultDir(folderDir, context);
-        }
-        file = new File(settingDir, folderDir);
-        if (!file.exists()) {
-            if (!file.mkdirs()) {
-                file = new File(settingDir + folderDir);
-                if (!file.exists()) {
-                    //noinspection ResultOfMethodCallIgnored
-                    file.mkdirs();
-                }
-            }
-        }
-        return file;
-    }
-
-    public static File getDownloadDir(Site site, Context context) {
-        File file;
-        SharedPreferences prefs = HentoidApplication.getAppPreferences();
-        String settingDir = prefs.getString(Constants.SETTINGS_FOLDER, "");
-        String folderDir = site.getFolder();
-        if (settingDir.isEmpty()) {
-            return getDefaultDir(folderDir, context);
-        }
-        file = new File(settingDir, folderDir);
-        if (!file.exists()) {
-            if (!file.mkdirs()) {
-                file = new File(settingDir + folderDir);
-                if (!file.exists()) {
-                    //noinspection ResultOfMethodCallIgnored
-                    file.mkdirs();
-                }
-            }
-        }
-        return file;
-    }
-
-    public static File getDefaultDir(String dir, Context context) {
-        File file;
-        try {
-            file = new File(Environment.getExternalStorageDirectory()
-                    + "/" + Constants.DEFAULT_LOCAL_DIRECTORY + "/" + dir);
-        } catch (Exception e) {
-            file = context.getDir("", Context.MODE_WORLD_WRITEABLE);
-            file = new File(file, "/" + Constants.DEFAULT_LOCAL_DIRECTORY);
-        }
-
-        if (!file.exists()) {
-            if (!file.mkdirs()) {
-                file = context.getDir("", Context.MODE_WORLD_WRITEABLE);
-                file = new File(file, "/" + Constants.DEFAULT_LOCAL_DIRECTORY + "/" + dir);
-                if (!file.exists()) {
-                    //noinspection ResultOfMethodCallIgnored
-                    file.mkdirs();
-                }
-            }
-        }
-        return file;
     }
 
     public static <K> void saveJson(K object, File dir)
